refactor naming of the interface and its impl
