+++ lib/networking/category_api.dart
@@ -1,17 +1,9 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:kortobaa_core_package/kortobaa_core_package.dart';
 import 'package:riverpod_files/networking/api_constants.dart';
 
-abstract class ICategoryApi {
-  static final provider = Provider<ICategoryApi>((ref) {
-   
+import '../blocs/interfaces/category_interface.dart';
 
-       return CategoryApiImpl(remoteClient: ref.watch(RemoteClient.provider(null)));
 
-  });
-
-  Future<List<String>> fetchCategories();
-}
 
 class CategoryApiImpl implements ICategoryApi {
   final RemoteClient remoteClient;
