+++ lib/networking/products_api.dart
@@ -1,20 +1,15 @@
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:kortobaa_core_package/kortobaa_core_package.dart';
 import 'package:riverpod_files/networking/api_constants.dart';
 import 'package:riverpod_files/blocs/models/products_model/product_response.dart';
 
-abstract class INewApi {
-  static final provider = Provider<INewApi>((ref) {
-    return NewApiImpl(remoteClient: ref.watch(RemoteClient.provider(null)));
-  });
+import '../blocs/interfaces/products_interface.dart';
 
-  Future<ProductResponse> getProducts(int skip, int limit);
-}
 
-class NewApiImpl implements INewApi {
+
+class ProductsApiImpl implements IProductsApi {
   final RemoteClient remoteClient;
 
-  NewApiImpl({required this.remoteClient});
+  ProductsApiImpl({required this.remoteClient});
 
   @override
   Future<ProductResponse> getProducts(int skip, int limit) async {
